+++ src/controllers.rs
@@ -1,2 +1,4 @@
 pub mod index;
 pub mod puzzle;
+pub mod about;
+pub mod api;