+++ src/db/card.rs
@@ -5,7 +5,9 @@ use sqlx::Row;
 use sqlx::sqlite::SqliteRow;
 
 use crate::srs::{Card, Difficulty};
-use crate::db::{PuzzleDatabase, DbResult, Puzzle};
+use crate::db::{PuzzleDatabase, DbResult, Puzzle, ErrorDetails};
+
+use super::DatabaseError;
 
 /// A review record from the db.
 #[derive(Debug, Clone)]
@@ -14,6 +16,16 @@ pub struct Review {
     pub puzzle_id: String,
     pub difficulty: Difficulty,
     pub date: DateTime<FixedOffset>,
+    pub user_rating: Option<i64>,
+}
+
+/// A bucket of review scores for puzzles in the given review range with the given score.
+#[derive(Debug)]
+pub struct ReviewScoreBucket {
+    pub puzzle_rating_min: i64,
+    pub puzzle_rating_max: i64,
+    pub difficulty: Difficulty,
+    pub review_count: i64,
 }
 
 impl<'r> sqlx::FromRow<'r, SqliteRow> for Review
@@ -32,14 +44,16 @@ impl<'r> sqlx::FromRow<'r, SqliteRow> for Review
                     index: "date".to_string(),
                     source: e.to_string().into(),
                 })?,
+            user_rating: row.try_get("user_rating").ok(),
         })
     }
 }
 
-impl<'r> sqlx::FromRow<'r, SqliteRow> for Card
-{
-    fn from_row(row: &'r SqliteRow) -> Result<Self, sqlx::Error> {
-        Ok(Self {
+impl PuzzleDatabase {
+    /// Build a card from a result row. The reason we have it defined here instead of as a FromRow
+    /// instance is because we need access to self.srs_config.
+    fn card_from_row<'r>(&self, row: &'r SqliteRow) -> Result<Card, sqlx::Error> {
+        Ok(Card {
             id: row.try_get("puzzle_id")?,
             interval: Duration::seconds(row.try_get("interval")?),
             review_count: row.try_get("review_count")?,
@@ -50,11 +64,10 @@ impl<'r> sqlx::FromRow<'r, SqliteRow> for Card
                     index: "due".to_string(),
                     source: e.to_string().into(),
                 })?,
+            srs_config: self.srs_config,
         })
     }
-}
 
-impl PuzzleDatabase {
     /// Get the next due review. min_interval allows us to filter out cards with short intervals
     /// (e.g. because they're still in learning), because otherwise they'll show up, possibly
     /// repeatedly if learning or relearning, before other cards that are due later today.
@@ -84,7 +97,7 @@ impl PuzzleDatabase {
             .fetch_optional(&self.pool)
             .await?
             .map(|row: SqliteRow| {
-                let card: Card = sqlx::FromRow::from_row(&row)?;
+                let card: Card = self.card_from_row(&row)?;
                 let puzzle: Puzzle = sqlx::FromRow::from_row(&row)?;
                 Ok((card, puzzle))
             })
@@ -95,7 +108,7 @@ impl PuzzleDatabase {
     pub async fn get_card_by_id(&self, puzzle_id: &str) -> DbResult<Option<Card>> {
         log::info!("Getting card for puzzle {puzzle_id}");
 
-        let query = sqlx::query_as("
+        let query = sqlx::query("
             SELECT *
             FROM cards
             WHERE puzzle_id = ?
@@ -103,8 +116,10 @@ impl PuzzleDatabase {
 
         Ok(query
             .bind(puzzle_id)
+            .map(|row| self.card_from_row(&row))
             .fetch_optional(&self.pool)
-            .await?)
+            .await?
+            .transpose()?)
     }
 
     /// Update (or create) a card by ID.
@@ -133,8 +148,8 @@ impl PuzzleDatabase {
     pub async fn add_review_for_user(&mut self, review: Review) -> DbResult<()>
     {
         let query = sqlx::query("
-            INSERT INTO reviews (user_id, puzzle_id, difficulty, date)
-            VALUES (?, ?, ?, ?)
+            INSERT INTO reviews (user_id, puzzle_id, difficulty, date, user_rating)
+            VALUES (?, ?, ?, ?, ?)
         ");
 
         query
@@ -142,35 +157,41 @@ impl PuzzleDatabase {
             .bind(&review.puzzle_id)
             .bind(review.difficulty.to_i64())
             .bind(review.date.to_rfc3339())
+            .bind(review.user_rating)
             .execute(&self.pool)
             .await?;
 
         Ok(())
     }
 
-    /// Get up to the last n reviews for a user, and the rating for each one.
-    pub async fn last_n_reviews_for_user(&self, user_id: &str, max_reviews: i64)
-        -> DbResult<Vec<(Review, i64)>>
+    /// Get a rating history for a user. 
+    pub async fn get_user_rating_history(&self, user_id: &str)
+        -> DbResult<Vec<(DateTime<FixedOffset>, i64)>>
     {
-        let query = sqlx::query("
-            SELECT reviews.*, puzzles.rating
-            FROM reviews
-            INNER JOIN puzzles ON reviews.puzzle_id = puzzles.puzzle_id
-            WHERE reviews.user_id = ?
-            AND puzzles.rating NOT NULL
-            ORDER BY date DESC
-            LIMIT ?
-        ");
+        let query = sqlx::query(r#"
+            SELECT * FROM (
+                SELECT reviews.date, max(user_rating) as max_rating
+                FROM reviews
+                WHERE user_id = ?
+                AND user_rating IS NOT NULL
+                -- Group by day and then hour
+                GROUP BY date(reviews.date), strftime('%H', reviews.date)
+                -- Add the user's current rating to the end
+                UNION SELECT strftime("%Y-%m-%dT%H:%M:%SZ", datetime("now")),
+                    rating FROM users WHERE id = ?
+            )
+            ORDER BY datetime(date)
+        "#);
 
         Ok(query
             .bind(user_id)
-            .bind(max_reviews)
+            .bind(user_id)
             .fetch(&self.pool)
             .map(|row: Result<SqliteRow, _>| {
                 let row = row?;
-                let review: Review = sqlx::FromRow::from_row(&row)?;
-                let rating: i64 = row.try_get("rating")?;
-                Ok((review, rating)) as Result<_, sqlx::Error>
+                let date = DateTime::parse_from_rfc3339(row.try_get("date")?)?;
+                let rating: i64 = row.try_get("max_rating")?;
+                Ok((date, rating)) as Result<_, DatabaseError>
             })
             .try_collect()
             .await?)
@@ -252,5 +273,45 @@ impl PuzzleDatabase {
            .await?
            .map(|row| row.try_get("card_count"))
            .unwrap_or(Ok(0))?)
+    }
+
+    /// Get the review score history for a user, in buckets of `rating_bucket_span` rating span.
+    /// e.g. rating_bucket_span = 50 means you'll get buckets every 50 puzzle rating, so 450-500,
+    /// 500-550, etc.
+    pub async fn get_review_score_history(&self, user_id: &str, rating_bucket_span: i64)
+        -> DbResult<Vec<ReviewScoreBucket>>
+    {
+        let query = sqlx::query("
+            SELECT reviews.difficulty,
+                puzzles.rating - puzzles.rating % ? AS min_rating,
+                count(reviews.difficulty) AS review_count
+            FROM reviews
+            JOIN puzzles ON reviews.puzzle_id = puzzles.puzzle_id
+            WHERE reviews.user_id = ?
+            GROUP BY min_rating, difficulty
+            ORDER BY min_rating, difficulty
+        ");
+
+        Ok(query
+            .bind(rating_bucket_span)
+            .bind(user_id)
+            .fetch(&self.pool)
+            .map(|row| {
+                let row = row?;
+                let min_rating = row.try_get("min_rating")?;
+                Ok(ReviewScoreBucket {
+                    puzzle_rating_min: min_rating,
+                    puzzle_rating_max: min_rating + rating_bucket_span,
+                    difficulty: Difficulty::from_i64(row.try_get("difficulty")?)
+                        .map_err(|e| DatabaseError::ParsingError(ErrorDetails {
+                            backend: "srs".to_string(),
+                            description: e.to_string(),
+                            source: Some(e),
+                        }))?,
+                    review_count: row.try_get("review_count")?,
+                }) as DbResult<ReviewScoreBucket>
+            })
+            .try_collect::<Vec<ReviewScoreBucket>>()
+            .await?)
     }
 }